Refactor GroupList & Groups
