--- static/js/containers/GroupListContainer.jsx
@@ -1,11 +1,0 @@
-import { connect } from 'react-redux';
-
-import GroupList from '../components/GroupList';
-
-const mapStateToProps = (state, ownProps) => (
-  {
-    groups: state.groups.latest
-  }
-);
-
-export default connect(mapStateToProps)(GroupList);
