--- static/js/containers/Groups.jsx
@@ -1,12 +1,0 @@
-import { connect } from 'react-redux';
-
-import Groups from '../components/Groups';
-
-
-const mapStateToProps = (state, ownProps) => (
-  {
-    profile: state.profile
-  }
-);
-
-export default connect(mapStateToProps)(Groups);
